requirements.txt
pandas~=1.0.5
yfinance~=0.1.54
matplotlib~=3.2.2
seaborn~=0.10.1
mysql-connector-python~=8.0.20
PyMySQL~=0.9.3
SQLAlchemy~=1.3.18
numpy~=1.19.0
sympy~=1.6.1
scikit-learn~=0.23.1